Move gui building to different class
